Rearrange some tasks
We have a gitolite/check_fedmsg_hooks role, which installs a script and
schedules it.

Turns out, this script does more than just checking the fedmsg hooks,
depending on the command-line arguments used when running it.

As such, it makes sense to separate it out into its own role, and make
the gitolite/check_fedmsg_hooks role (and any other one using the
script) depend on it.

For example, this script is used for Fedora Hosted (still in Puppet),
and will soon be used for a new distgit hook check.