Drop the -s/--source branch argument in mkbranch
In the process of automating the git branch creation based on fedmsg
message we can no longer rely on passing a -s/--source branch argument
to the mkbranch script.
As decided in https://fedorahosted.org/rel-eng/ticket/5931 the branches
will be created with the first commit of the master branch.

This commit thus drops the use of -s/--source and SRC_BRANCH in favor of
creating the branch with the first commit of the master branch:
git branch --no-track $BRANCH `git rev-list --max-parents=0 master`