Move back the variables to what they should be
fasClient was failing as it could not put the username in places.
At first I thought this was an issue related to the first time we run fasClient
and I thus introduced a way-around by making the first run be w/o the use of
these variables and later in the role, add a new fas.conf template that would
use these variables.
It turned out that the problem was coming from fasClient itself (change merged
upstream but not deployed) so all my efforts were useless.

Reverting all things back to their original state and we'll just look into
fixing fasClient properly.