add rsyslog to global_pkgs_inst
As part of
  http://fedoraproject.org/wiki/Changes/NoDefaultSyslog
rsyslog is no more installed by default since F20+.

role/base refer to rsyslog on several places: enable that service, start
it... but never install it (only in Koji builder playbook install it,
but not as part of the role).

Which cause our playbooks to fail on F21 machines.

Allowed in Freeze by:
https://lists.fedoraproject.org/pipermail/infrastructure/2015-March/015840.html