Block the Riddler.io bot from accessing bodhi
This bot does NOT follow robots.txt, even though it announces that
it does, and it hits Bodhi so much that it pulls the bodhi servers
under.

An email has been sent to the maintainer of the bot, but for the
time being, let's block it from using any bodhi resources.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>