Distribute the hosts to be crawled by MM2 between the existing crawlers
This commit adds two new scripts to the crawlers:
 * mm2_get-highest-active-host-id: this queries the MM database and returns
   the id of the active mirror with the highest id
 * run_crawler.sh: this calculates the right startid and stopid parameters
   depending on the number of available crawlers

These scripts are integrated into the start of the crawler so that each
crawler only crawls a subset of the existing mirrors. This distribution
is not perfect as it pretends that the active mirrors are equally
distributed under the total number of existing mirrors.