Remember the start time of the last umdl run
The script which checks if umdl should be run used to store
the time of the end of the current umdl run. It actually should
remember the time of the start of this run as we want to make sure
to pick up all changes since the start of the current run.

Also print the date and time in the same line as the logging.