Increase crawler timeout from 2h to 3h.
Since MM2 is in production about 140 mirrors have been auto-disabled
due to crawler timing out after 2 hours (default). Try if it works
better with 3 hours. This in combination with the previous commit
to decrease the load on the database should help to auto disable
less good mirrors. Especially mirrors who mirroring almost
everything can hardly be crawled within the 2 hour limit. Unfortunately
the limit is per host and not category.