Increase cache time for mirrorlist
The mirrorlist which can be viewed in the browser used to be generated
once or twice per day with MM1. This is listed is now cached in varnish
but only for two minutes. This patch increases the cache time to 6 hours.
Also increase cache time for the static mirrormanager content.