Decrease vfs_cache_pressure on mm backend systems
To decrease the crawl duration on the mirrors we have been
recommending to lower the default value of vfs_cache_pressure
from 100 to 10. This causes the kernel to prefer to keep dentries
when under memory pressure. Let's also set it on the mm backend
systems as umdl is mainly looking at the metadata.