Better distribution of mirrors to crawl.
The script mm2_get-highest-active-host-id used to return the highest
ID of the active mirrors. This number was divided by the number of
active crawlers and then each crawler got its share of mirrors to crawl.
This did not take into account that more active mirrors are in the
higher IDs as old mirror IDs are not re-used and thus one crawler was
getting much more mirrors to crawl than another. The new script (which
will be renamed) now divides the list correctly by returning exactly the
fraction which each crawler should crawl.