Install python-GeoIP for the new crawler.
The latest crawler needs python-GeoIP for continent specific crawls.
This should be a dependency of the mirrormanager2-crawler package but as
it is not python-GeoIP has to be installed manually. For some reason the
package was installed on mm-crawler02 but not on mm-crawler01. All
mirrors that should have been crawled from mm-crawler01 have not been
crawled for the last few days.