Handle the fact that the static files are in two locations
The static files are split over two places:
- /usr/share/pdc/static/ which contains the files specific to pdc
- /usr/share/patternfly1/resources/ which contains the files from patternfly1

This means that we much teach apache to go to both places to look for the static
files. This is quite annoying and thankfully something that Patrick Uiterwijk
and I already looked into for flask-multistatic.
So the solution was already known and documented:
https://pagure.io/flask-multistatic/blob/master/f/example.conf