openqa workers: install UEFI firmware package
This comes from the kraxel.org repo and was added to the infra
repo by nirik. Note that it is technically non-freely licensed.
It contains a FAT implementation; Microsoft has granted an
exemption from the FAT patent purely for the purpose of
implementing a UEFI firmware, so there is a use restriction on
this code. License can be found (as I write this) at:

https://github.com/tianocore/edk2/blob/master/FatBinPkg/License.txt