Don't do updates in playbooks (for a number of reasons).
1. We want all playbooks to be idempotent. If we add things like this
it means the playbook will change every single time we run it.

2. Things like this mean that if we need to change one config or
something to get the service working we may have to wait a long time
for it to apply tons of pending updates, which may break more things.

3. On all Fedora machines (that have base role anyhow) we already run
dnf-automatic to apply all security updates. We could expand that to
apply all updates if you like. Doing them daily is much better for
finding regressions

4. We expect updates are applied normally by dnf-automatic or by
sysadmins so we can know whats in the updates and be ready for issues
or restarting things.