beaker/server: configure bkr client pointing back to localhost
We can use this for registering and configuring the lab controller as
part of the deployment playbooks. It will probably come in handy for
other stuff as well, like ad hoc tasks and testing.

The template is intentionally not named
templates/root/.beaker_client/config.j2 following the same pattern as
the other templates, because it seemed like a bad idea to have
a template inside a hidden directory.