add an 'ansible_ifcfg_whitelist' feature and use it for openqa
semi-acked by nirik (but he'll deny it furiously and it's all
my fault if everything blows up): for openQA's openvswitch
stuff I need a 'br0' and a 'tap0' that I don't want the base
role to mess with, but I *do* want the base role to configure
eth0 for me. ansible_ifcfg_blacklist isn't granular enough. So
let's invent ansible_ifcfg_whitelist, which if defined is a
list of interface names you want the base role to configure.
Any interface not in the list is left alone.