Fix OpenShift key sync
Turns out the used library only works with a specific API version
but it didn't specify that unless the user tells it to.
This pins us on openshift API 1.5 until support for that is dropped.

This also fixes the issue where we pass the key arguments as keyword
arguments but the key_add function did not expect them as such.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>