tune copr-dist-git.conf
cgit can take up to 10 minutes - rarely even longer
Once run, it will create cache, but initial run is long.
So I increased timeout a lot.
But to not take down the server with lots of processes, I limited concurent processes.
In normal situation it is more than enough.
And when cgit creates cache, it can lock down few users. But it will create the cache and next run will be fast.