mm2_crawler: make sure only one crawler is running
To make sure only one cron started crawler is running the previous
running (cron) crawlers are being signaled to shut down. The crawler can
try to gracefully shutdown if it gets the signal SIGALRM(14). After the
signal we wait for 5 minutes to give the crawler a chance to shutdown.
After that the crawler is killed. To make sure we only end the cron
started crawler we look for the following process "/usr/bin/python
/usr/bin/mm2_crawler --threads 25".

Signed-off-by: Adrian Reber <adrian@lisas.de>