Add a root logger to the fedmsg base logging config
This alters the logging configuration for all fedmsg applications. It
adds a root logger configuration which handles messages that are not
covered by other loggers (and which also have ``propagate=False``). This
means that all Python modules will now be able to emit log messages,
rather than those in the ``fedmsg`` or ``moksha`` Python packages (or
other packages that use their loggers).

Signed-off-by: Jeremy Cline <jeremy@jcline.org>