Disable registry caching so we can avoid Redis.
Patrick and Adam and I had a meeting this morning to discuss how
to deploy Redis to be used as a shared cache for the registry
nodes. Patrick suggested that we could instead disable the regstry
caching layer and rely on Varnish to do all our caching, which
makes the system simpler. This commit disables the in-memory cache
of the registries so they won't get out of sync on write
operations.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>