Add a cron job for rawhide nightly with DNF
In order to compare YUM and DNF compose, this will start a simplified
compose (no extra deliverables, just the packages) and create logs with
differences. The compose starts 12 hours after the real rawhide, which
should prevent any negative effects on the production compose (while
it's possible for the nightly to not finish yet, after 12 hours it would
only be waiting on koji tasks and not needed local resources much).

See https://pagure.io/pungi-fedora/pull-request/178 for details.

Signed-off-by: Lubomír Sedlář <lsedlar@redhat.com>