add manage-container-images role, use w/ osbs-cluser
Currently the push-docker role requires a docker daemon to be
running on a machine somewhere which we don't want because it's
privileged, error prone, and only manages docker registry content.
This role instead uses skopeo[0], which is not privileged and
understands how to manage many types of OCI[1] compliant container
images.

[0] - https://github.com/projectatomic/skopeo
[1] - https://www.opencontainers.org/

Signed-off-by: Adam Miller <admiller@redhat.com>