rsync changes for syncing bits to inside RH
expose /mnt/koji/compose via rsync so that we can rsync the bits
inside of Red Hat, for testing as well as mirroring the release.
Its drops a terrible ssh hack and allows a path for automation.

allow the locations in RH infra that do mirroring of bits to use tier 1

Signed-off-by: Dennis Gilmore <ausil@fedoraproject.org>