Add a package owner email alias script to use Pagure instead of PkgDB and stop querying FAS in staging
As part of the Modularity project, we've decided that we need to have different package
branches with different lifecycles. To read more about this, please view this document:
https://fedoraproject.org/wiki/Infrastructure/Factory2/Focus/ArbitraryBranching. To enable
this, we've decided to rely on Pagure and PDC for package information instead of PkgDB.
This commit is an attempt to modify an existing script that queried PkgDB and FAS for packager
emails with one that uses the new architecture described above.

Also, instead of querying FAS for email information, we've decided to use the email alias
USER@fedoraproject.org. This will be a performance increase and protect us against FAS API
changes.