Create "owner-sync-pagure.j2" which is an eventual replacement for "owner-sync-pkgdb.j2".
With the eventual decommissioning of PkgDB, we need a way to get package ownership from Pagure
instead of dist-git when using the sync script, which sets ownership in Koji. This will
eventually be turned on in staging when Pagure is updated to include patches to the API
in master that this script requires.

For more information on the decommissioning of PkgDB, please read:
https://lists.fedoraproject.org/archives/list/infrastructure@lists.fedoraproject.org/thread/C5YIZ7RG7UW2AYS6Y5IQXV3UY44UW24Y/#C5YIZ7RG7UW2AYS6Y5IQXV3UY44UW24Y

This is part of the "Arbitrary Branching" effort. For information on this, please read:
https://fedoraproject.org/wiki/Infrastructure/Factory2/Focus/ArbitraryBranching