bump repo_tasks_limit to allow an increase in repos
repo_tasks_limit defaults to 10, it sets the maximum repo tasks in
total, combining both the rpm repos and unused in fedora maven
repos. max_repo_tasks in the config was set to 15 but in reality
we never had more than 10 due to the extra limitation. Both are
set to 15 now

Signed-off-by: Dennis Gilmore <ausil@fedoraproject.org>