mirrormanager2: change umdl-required script
The umdl-required script used to query the message bus if there have
been changes on the master mirror. With the new stat()-less umdl
implementation the message bus is no longer queried but only the ctime
of the fullfiletimelist-* is checked to see if it has changed since the
last run.

Signed-off-by: Adrian Reber <adrian@lisas.de>