taskotron-client: replace vars in yumrepoinfo.conf
Instead of keeping modified version of the config file and merging the
changes every time (which is very error prone - I broke it in
cddf101acb9d), rather store the vanilla version and perform
modifications on top of it. These modifications will fail if the keys
change, hopefully reducing silent failures.