openqa/worker: add boot script to fix KVM perms and disable SMT
For some reason /dev/kvm has 0600 perms after boot on the ppc64
worker host. Also, qemu won't run unless SMT is turned off, on
ppc64. I've just been doing this manually every time the box got
restarted, but that's dumb, so let's make it happen on boot with
a script and a service to run it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>