mm-propagation: use pdc instead of pkgdb for active releases
The propagation scripts were still using pkgdb to get a list of active
releases. This switches the scripts to use pdc instead of pkgdb. One
difference between pdc and pkgdb is that pkgdb did not mark a release
as active until the actual release day. In pdc the release (in this case
f27) is already marked as active and it is not clear how to filter it
out until the official release.

As the propagation scripts are looking at the updates-released repository
it will not provide useful data for f27 until content is pushed to the
updates-released repository.

Signed-off-by: Adrian Reber <adrian@lisas.de>