mm2-crawler: reduce database load during crawling
The two crawlers used to start 25 threads (each) every 12 hours to crawl
the mirrors. The second crawler was already started two hours later to
reduce the database load. This commit increases the delay of the second
crawler to 6 hours, so that every crawler has the MM database for 6
hours on its own. At the same time the number of parallel crawlers is
reduced from 25 to 20 to also reduce the load on the database.

In addition the crawl timeout has been increased from 3 to 4 hours. This
is related to the fact the especially pub/archive has grown and
pub/fedora with the addition of the modular tree. Crawl timeouts can now
be seen more often, which can lead to mirrors being auto-disabled.

The main reason for these changes is that it can be seen in the logs that
the actual crawling of the mirrors does not always require most of time
but updating the state of all directories of each mirror in the database
can take a very long time. By reducing the number of parallel accesses to
the database, in the best case from 50 to 20, the crawling should get
faster (hopefully).

Signed-off-by: Adrian Reber <adrian@lisas.de>