Remove osbs-orchestrator-cluster.yml; align stg/prod OSBS
For the time being, this needs to be removed. This work has been
put on hold until a new Fedora Engineering Team member is onboarded
in order to take this work over and in the mean time there's no
sense in leaving stage OSBS broken for users.

This commit also brings stage OSBS back into alignment with the
production OSBS which is the "old" OSBS Architecture as defined in
the upstream documentation:

    https://osbs.readthedocs.io/en/latest/multiarch.html

Signed-off-by: Adam Miller <admiller@redhat.com>