wikitcms token: create directory, install for fedmsg and root
Previous commit didn't work as the directory doesn't exist.
Also, I think we need the file in two places, one for root
(for if an admin runs something wikitcms-ish manually while
logged in as root) and one for fedmsg (for the usual case
when the wiki is updated by fedmsg consumers). So do that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>