git (post-receive-alternativearch): use unicode for git output
The commit message and diff is stored in `full_change` as a string, via
`read_output()`.  This is passed to `TEXT` which is a unicode string.

When a commit message or diff contains non-ascii characters we get:

    UnicodeDecodeError: 'ascii' codec can't decode byte ...

Encode git output returned from `read_output()` as a unicode string and
define `full_change` as unicode for completeness.

Fixes #6040.