Allow openQA staging to publish 'ci' fedmsgs
This is part of a Big Secret Red Hat Conspiracy. Look away now!
OK, we're working on a standardized messaging format for testing
related messages, and this is part of testing out how that looks
with a Fedora testing system. Won't go to prod till it's a bit
more locked in.

Signed-off-by: Adam Williamson <awilliam@redhat.com>