Bump openQA asset size limits for staging
This is due to an upstream change in how the asset reduction
logic works: it now includes 'fixed' assets in the total size
calculation, so we should include the size of those assets
(which is currently ~110GB) in our limits. This is only on stg
for now as only stg has been updated to the new upstream code.

Signed-off-by: Adam Williamson <awilliam@redhat.com>