openQA: tweak ansible_ifcfg_whitelist handling for workers
As we have more and more tap workers, it's no longer practical
to do this for the whole group, as each one seems to have a
*different* set of interfaces a) present and b) 'active'. So
let's just take it out of the group vars and do it per-host
instead. Each tap worker only actually has *one* real hardware
interface that should be handled by network / NetworkManager,
so we just list that one per-host (it's eth0 for all of them
except the ppc64 host, which is eth2 for some damn reason).

Signed-off-by: Adam Williamson <awilliam@redhat.com>