mm2_crawler: reduce number of threads and increase timeout
This reduce the number of threads used per category by one for each
category. mm-crawler02 sometimes has to OOM kill one of the crawlers and
by using less crawl threads the crawler should not OOM any more. This
needs to be monitored to find the optimal number of threads.

This also increases the timeout for the archive crawling from 4 hours to
5 hours.

Signed-off-by: Adrian Reber <adrian@lisas.de>