waiverdb: trigger import-image instead of start-build
Nowadays we have an imagestream, which pulls pre-built waiverdb images
from quay.io, and a buildconfig, which then layers some
Fedora-infra-specific stuff on top.

When we have a new version to deploy, the imagestream should pick up the
new image from quay.io, which triggers a build of the buildconfig, which
triggers a deploy of the deploymentconfig. So we shouldn't need to
explicitly fire off start-build anymore.

We *shouldn't* need to explicitly fire off import-image either, because
it is set to "scheduled" pull policy. But I am not sure what the
schedule is. Right now it says it last polled 5 days ago... that may
itself be a bug. Regardless, this gives us a nice way to make it pull
the latest image.

This patch adds a new, very simple role "import-image" to fire off the
oc import-image command, following the precedent of the existing
"start-build" role.