greenwave: add subject_type key to policies for upcoming 0.8 release
In Greenwave 0.8 (not released yet) subject_type will replace
relevance_key and relevance_value:

https://docs.pagure.org/greenwave/release-notes.html#next-release

Add the appropriate subject_type keys to all policies so that we can
roll out the latest code to staging.

This is not wrapped in a conditional because older versions of Greenwave
will just ignore the new subject_type key, while newer versions of
Greenwave will ignore the relevance_key and relevance_value keys.

Later, once Greenwave 0.8 is released and rolled out in stage and prod,
we can remove the relevance_key and relevance_value from the policies.