bodhi-pungi: specify ostree ref in config and updates for f29
Now that we can specify the ostree_ref in the config we can stop
having to choose various fedora-atomic-host-updates-{stable,testing}.json
files.

For silverblue we are also changing the ref to silverblue away from
workstation so we'll only specify ostree_ref if f29+ and let the json
files still specify "workstation" for f27 and f28.