Turn off fedmsg on ci-cc
This is done for two reasons:
a) We do not really care about sending fedmsg messages from this resultsdb
instance, they should rather be sent via the official one once they have
been synced accross.
b) Installing the base fedmsg role leads to a huge number of hosts that
fedmsg will try to listen to, in fact, it results in too many hosts that
fedmsg will try to connect to, leading to the core app of this host
(resultsdb-listener) to not listen to the host it is designed to listen
to (ci.centos.org) making this service effectively not working.

So by removing the base fedmsg role we'll ensure fedmsg only connects
to the hosts we care about and by removing resultsdb-fedmsg we won't try
to have resultsdb send fedmsg messages upon receiving new test results.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>